+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.6.3 - Unreleased
 
+### Added
+* `equal_with_nan` in `util_numpy`
 
 ## Version 0.6.2 - Released 2022-06-15